+++ relay-dynamic-config/src/global.rs
@@ -181,6 +181,26 @@ pub struct Options {
     )]
     pub sessions_eap_rollout_rate: f32,
 
+    /// Rollout rate for accepted outcomes being emitted by EAP instead of Relay.
+    ///
+    /// Rate needs to be between `0.0` and `1.0`.
+    #[serde(
+        rename = "relay.eap-outcomes.rollout-rate",
+        deserialize_with = "default_on_error",
+        skip_serializing_if = "is_default"
+    )]
+    pub eap_outcomes_rollout_rate: f32,
+
+    /// Rollout rate for accepted outcomes for spans being emitted by EAP instead of Relay.
+    ///
+    /// Rate needs to be between `0.0` and `1.0`.
+    #[serde(
+        rename = "relay.eap-span-outcomes.rollout-rate",
+        deserialize_with = "default_on_error",
+        skip_serializing_if = "is_default"
+    )]
+    pub eap_span_outcomes_rollout_rate: f32,
+
     /// All other unknown options.
     #[serde(flatten)]
     other: HashMap<String, Value>,